fix(overlay): using hostData for zIndex
