+++ packages/core/src/components/menu/menu.tsx
@@ -237,8 +237,10 @@ export class Menu {
       return Promise.resolve();
     }
     // Destroy existing animation
-    this.animation && this.animation.destroy();
-    this.animation = null;
+    if (this.animation) {
+      this.animation.destroy();
+      this.animation = null;
+    }
     this.width = width;
 
     // Create new animation
