+++ packages/core/src/utils/helpers.ts
@@ -15,16 +15,6 @@ export function assert(actual: any, reason: string) {
   }
 }
 
-export function autoFocus(containerEl: HTMLElement): HTMLElement|null {
-  const focusableEls = containerEl.querySelectorAll('a[href], area[href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), button:not([disabled]), [tabindex="0"]');
-  if (focusableEls.length > 0) {
-    const el = focusableEls[0] as HTMLInputElement;
-    el.focus();
-    return el;
-  }
-  return null;
-}
-
 export function now(ev: UIEvent) {
   return ev.timeStamp || Date.now();
 }
